+++ massiv/src/Data/Massiv/Array/Manifest.hs
@@ -51,7 +51,6 @@ module Data.Massiv.Array.Manifest
   , toStorableMVector
   -- ** Direct Pointer Access
   , withPtr
-  , unsafeWithPtr
   -- * Unboxed
   , U(..)
   , Unbox
@@ -60,11 +59,14 @@ module Data.Massiv.Array.Manifest
   , toUnboxedMVector
   -- * ByteString Conversion
   , fromByteString
+  , castFromByteString
   , toByteString
+  , castToByteString
   , toBuilder
   ) where
 
 import           Data.ByteString                      as S
+import           Data.ByteString.Internal
 import           Data.ByteString.Builder
 import           Data.ByteString.Unsafe               as SU
 import           Data.Massiv.Array.Manifest.Boxed
@@ -90,24 +92,38 @@ fromByteString ::
 fromByteString comp bs = MArray comp (SafeSz (S.length bs)) (SU.unsafeIndex bs)
 {-# INLINE fromByteString #-}
 
--- | /O(n)/ - For now only sequentially convert an array into a strict ByteString
+-- | /O(n)/ - Convert any source array into a strict `ByteString`. In case when the source array is
+-- actually storable, no memory copy will occur.
 --
 -- @since 0.2.1
 toByteString ::
      Source r ix Word8
   => Array r ix Word8 -- ^ Source array
   -> ByteString
-toByteString arr =
-  fst $ unfoldrN (totalElem (size arr)) (\ !i -> Just (unsafeLinearIndex arr i, i + 1)) 0
+toByteString = castToByteString . convert
+  --fst $ unfoldrN (totalElem (size arr)) (\ !i -> Just (unsafeLinearIndex arr i, i + 1)) 0
 {-# INLINE toByteString #-}
 
--- | /O(n)/ - Conversion of array monoidally into a string Builder.
+-- | /O(n)/ - Conversion of array monoidally into a ByteString Builder.
 --
 -- @since 0.2.1
 toBuilder :: Source r ix e => (e -> Builder) -> Array r ix e -> Builder
-toBuilder f = foldMono f
+toBuilder = foldMono
 {-# INLINE toBuilder #-}
 
+-- | /O(1)/ - Cast a `S`torable array into a strict `ByteString`
+--
+-- @since 0.3.0
+castToByteString :: Array S ix Word8 -> ByteString
+castToByteString = (\(fp, len) -> PS fp 0 len) . unsafeArrayToForeignPtr
+{-# INLINE castToByteString #-}
+
+-- | /O(1)/ - Cast a strict `ByteString` into  a`S`torable array
+--
+-- @since 0.3.0
+castFromByteString :: Comp -> ByteString -> Array S Ix1 Word8
+castFromByteString comp (PS fp offset len) = unsafeArrayFromForeignPtr comp fp offset (Sz len)
+{-# INLINE castFromByteString #-}
 
 -- $boxed_conversion_note
 --