+++ massiv/tests/Main.hs
@@ -3,7 +3,6 @@ module Main where
 import System.IO (BufferMode(LineBuffering), hSetBuffering, stdout)
 import Test.Hspec
 import Spec
-import Data.Massiv.Array
 
 main :: IO ()
 main = do